refactor: process_repo -> process_sorries
